Fix change requests
